update checker moved to another file
